♻ refactor: melhora ui dos counters de actions e altera componentes de pasta
