+++ src/App.tsx
@@ -5,15 +5,15 @@ import Login from './pages/Login';
 import Register from './pages/Register';
 import ForgotPassword from './pages/ForgotPassword';
 import ResetPassword from './pages/ResetPassword';
-import AuthInitializer from './components/AuthInitializer';
+import AuthInitializer from './components/auth/AuthInitializer';
 import CustomToastContainer from './components/ui/CustomToastContainer';
-import { PrivateRoute } from './components/PrivateRoute';
-import Contact from './components/Contact';
+import { PrivateRoute } from './components/auth/PrivateRoute';
+import Contact from './components/home/Contact';
 import Feed from './pages/Feed';
 import ProfileView from './pages/ProfileView';
 import ProfileEdit from './pages/ProfileEdit';
 import AccountSettings from './pages/AccountSettings';
-import CreatePostPage from './pages/posts/CreatePostPage';
+import CreatePostPage from './pages/CreatePostPage';
 import EventsPage from './pages/EventsPage';
 import DonationsPage from './pages/DonationsPage';
 import ServicesPage from './pages/ServicesPage';
