+++ src/components/home/Events.tsx
@@ -2,23 +2,24 @@ import { Swiper, SwiperSlide } from 'swiper/react';
 import 'swiper/css';
 import 'swiper/css/pagination';
 import { Pagination, Navigation } from 'swiper/modules';
-import Typography from './ui/Typography';
-import CardButton from './ui/CardButton';
-import { useHomeServices } from '../hooks/useHomeData';
-import { resolveImageUrl } from '../utils/resolveImageUrl';
+import Typography from '../ui/Typography';
+import CardButton from '../ui/CardButton';
+import { useHomeEvents } from '../../hooks/useHomeData';
+import { resolveImageUrl } from '../../utils/resolveImageUrl';
+import formatDateBR from '../../utils/formatDateBR';
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { useAuth } from '../hooks/useAuth';
+import { useAuth } from '../../hooks/useAuth';
 import { toast } from 'react-toastify';
-import { useScrollStore } from '../stores/scrollStore';
+import { useScrollStore } from '../../stores/scrollStore';
 
-export default function Services() {
-  const { services, loading, error } = useHomeServices(6);
+export default function Eventos() {
+  const { events, loading, error } = useHomeEvents(6);
   const [imageErrors, setImageErrors] = useState<{ [key: string]: boolean }>(
     {}
   );
   const navigate = useNavigate();
-  const { isAuthenticated } = useAuth(); // Usar o hook de autenticação
+  const { isAuthenticated } = useAuth(); // ✅ Adicionar hook de autenticação
 
   const handleImageError = (id: string) => {
     setImageErrors((prev) => ({ ...prev, [id]: true }));
@@ -26,12 +27,12 @@ export default function Services() {
 
   const setScrollTarget = useScrollStore((s) => s.setScrollTarget);
 
-  const handleServiceClick = (service: any) => {
-    const postId = service.postId || service.id;
+  const handleEventClick = (event: any) => {
+    const postId = event.postId || event.id;
     const targetUrl = `/post/${postId}`;
 
     if (!isAuthenticated) {
-      toast.info('Faça login para ver os detalhes do serviço');
+      toast.info('Faça login para ver os detalhes do evento');
       navigate('/login', {
         state: {
           from: targetUrl,
@@ -41,7 +42,7 @@ export default function Services() {
     }
 
     // grava na store QUAL seção e qual posição de scroll (apenas aqui)
-    setScrollTarget('services', window.scrollY);
+    setScrollTarget('events', window.scrollY);
 
     // navega para detalhes normalmente
     navigate(targetUrl);
@@ -50,7 +51,7 @@ export default function Services() {
   if (loading) {
     return (
       <section
-        id="services"
+        id="events"
         className="w-full py-12 px-4 md:px-8 bg-gray-50 mb-6"
       >
         <div className="max-w-3xl mx-auto text-center">
@@ -58,10 +59,10 @@ export default function Services() {
             variant="h1"
             className="text-3xl md:text-5xl font-bold text-primary"
           >
-            Serviços Disponíveis
+            Eventos Disponíveis
           </Typography>
           <Typography variant="p" className="text-gray-600 mt-2">
-            Carregando serviços...
+            Carregando eventos...
           </Typography>
         </div>
       </section>
@@ -71,7 +72,7 @@ export default function Services() {
   if (error) {
     return (
       <section
-        id="services"
+        id="events"
         className="w-full py-12 px-4 md:px-8 bg-gray-50 mb-6"
       >
         <div className="max-w-3xl mx-auto text-center">
@@ -79,7 +80,7 @@ export default function Services() {
             variant="h1"
             className="text-3xl md:text-5xl font-bold text-primary"
           >
-            Serviços Disponíveis
+            Eventos Disponíveis
           </Typography>
           <Typography variant="p" className="text-red-600 mt-2">
             {error}
@@ -89,10 +90,10 @@ export default function Services() {
     );
   }
 
-  if (services.length === 0) {
+  if (events.length === 0) {
     return (
       <section
-        id="services"
+        id="events"
         className="w-full py-12 px-4 md:px-8 bg-gray-50 mb-6"
       >
         <div className="max-w-3xl mx-auto text-center">
@@ -100,10 +101,10 @@ export default function Services() {
             variant="h1"
             className="text-3xl md:text-5xl font-bold text-primary"
           >
-            Serviços Disponíveis
+            Eventos Disponíveis
           </Typography>
           <Typography variant="p" className="text-gray-600 mt-2">
-            Nenhum serviço disponível no momento.
+            Nenhum evento disponível no momento.
           </Typography>
         </div>
       </section>
@@ -111,20 +112,17 @@ export default function Services() {
   }
 
   return (
-    <section
-      id="services"
-      className="w-full py-12 px-4 md:px-8 bg-gray-50 mb-6"
-    >
+    <section id="events" className="w-full py-12 px-4 md:px-8 bg-white mb-6">
       <div className="max-w-3xl mx-auto text-center">
         <Typography
           variant="h1"
           className="text-3xl md:text-5xl font-bold text-primary"
         >
-          Serviços Disponíveis
+          Eventos
         </Typography>
         <Typography variant="p" className="text-gray-600 mt-2">
-          Conheça os serviços oferecidos pela nossa rede social solidária,
-          desenvolvidos para apoiar e transformar vidas.
+          Participe e faça a diferença! Acompanhe nossos eventos e junte-se às
+          ações que transformam vidas.
         </Typography>
       </div>
 
@@ -153,51 +151,58 @@ export default function Services() {
             },
           }}
         >
-          {services.map((service) => (
-            <SwiperSlide key={service.id}>
+          {events.map((event) => (
+            <SwiperSlide key={event.id}>
               <div className="bg-white border rounded-lg shadow-md overflow-hidden flex flex-col h-full">
-                {/* Container da imagem com altura fixa mas proporção preservada */}
                 <div
                   className={`w-full aspect-[4/3] flex items-center justify-center rounded-t-lg overflow-hidden ${
-                    !service.image || imageErrors[String(service.id)]
+                    !event.image || imageErrors[String(event.id)]
                       ? 'bg-[#f0f9ff]'
                       : 'bg-gradient-to-b from-gray-50 to-gray-100'
                   }`}
                 >
                   <img
                     src={
-                      service.image && !imageErrors[String(service.id)]
-                        ? resolveImageUrl(service.image)
-                        : '/img/servico-solidario2.png'
+                      event.image && !imageErrors[String(event.id)]
+                        ? resolveImageUrl(event.image)
+                        : '/img/evento-solidario.png'
                     }
-                    alt={service.title}
+                    alt={event.title}
                     className="object-contain w-full h-full transition-transform duration-300"
-                    onError={() => handleImageError(String(service.id))}
+                    onError={() => handleImageError(String(event.id))}
                   />
                 </div>
 
-                {/* Resto do conteúdo permanece igual */}
+                {/* Conteúdo */}
                 <div className="p-3 flex flex-col flex-1">
                   <div className="h-10 mb-1">
                     <Typography
                       variant="h3"
                       className="text-base text-gray-900 line-clamp-2"
                     >
-                      {service.title}
+                      {event.title}
                     </Typography>
                   </div>
 
                   <div className="h-14 mb-2">
                     <Typography
                       variant="p"
-                      className="text-sm text-gray-600 line-clamp-2"
+                      className="text-sm text-gray-600 space-y-1"
                     >
-                      {service.description}
+                      <span>
+                        <strong>Data:</strong> {formatDateBR(event.date)}
+                      </span>
+                      <span className="flex">
+                        <strong className="flex-shrink-0">Local:</strong>
+                        <span className="line-clamp-2 ml-1">
+                          {event.location}
+                        </span>
+                      </span>
                     </Typography>
                   </div>
 
                   <div className="mt-2">
-                    <CardButton onClick={() => handleServiceClick(service)}>
+                    <CardButton onClick={() => handleEventClick(event)}>
                       Saiba Mais
                     </CardButton>
                   </div>
@@ -207,7 +212,7 @@ export default function Services() {
           ))}
         </Swiper>
 
-        {/* Botões de navegação - posicionados fora do card */}
+        {/* Botões de navegação */}
         <div className="swiper-button-prev text-3xl absolute left-0 top-1/2 transform -translate-y-1/2 text-gray-700 hover:text-blue-600"></div>
         <div className="swiper-button-next text-3xl absolute right-0 top-1/2 transform -translate-y-1/2 text-gray-700 hover:text-blue-600"></div>
       </div>
