+++ src/components/home/Services.tsx
@@ -2,24 +2,23 @@ import { Swiper, SwiperSlide } from 'swiper/react';
 import 'swiper/css';
 import 'swiper/css/pagination';
 import { Pagination, Navigation } from 'swiper/modules';
-import Typography from './ui/Typography';
-import CardButton from './ui/CardButton';
-import { useHomeEvents } from '../hooks/useHomeData';
-import { resolveImageUrl } from '../utils/resolveImageUrl';
-import formatDateBR from '../utils/formatDateBR';
+import Typography from '../ui/Typography';
+import CardButton from '../ui/CardButton';
+import { useHomeServices } from '../../hooks/useHomeData';
+import { resolveImageUrl } from '../../utils/resolveImageUrl';
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { useAuth } from '../hooks/useAuth';
+import { useAuth } from '../../hooks/useAuth';
 import { toast } from 'react-toastify';
-import { useScrollStore } from '../stores/scrollStore';
+import { useScrollStore } from '../../stores/scrollStore';
 
-export default function Eventos() {
-  const { events, loading, error } = useHomeEvents(6);
+export default function Services() {
+  const { services, loading, error } = useHomeServices(6);
   const [imageErrors, setImageErrors] = useState<{ [key: string]: boolean }>(
     {}
   );
   const navigate = useNavigate();
-  const { isAuthenticated } = useAuth(); // ✅ Adicionar hook de autenticação
+  const { isAuthenticated } = useAuth(); // Usar o hook de autenticação
 
   const handleImageError = (id: string) => {
     setImageErrors((prev) => ({ ...prev, [id]: true }));
@@ -27,12 +26,12 @@ export default function Eventos() {
 
   const setScrollTarget = useScrollStore((s) => s.setScrollTarget);
 
-  const handleEventClick = (event: any) => {
-    const postId = event.postId || event.id;
+  const handleServiceClick = (service: any) => {
+    const postId = service.postId || service.id;
     const targetUrl = `/post/${postId}`;
 
     if (!isAuthenticated) {
-      toast.info('Faça login para ver os detalhes do evento');
+      toast.info('Faça login para ver os detalhes do serviço');
       navigate('/login', {
         state: {
           from: targetUrl,
@@ -42,7 +41,7 @@ export default function Eventos() {
     }
 
     // grava na store QUAL seção e qual posição de scroll (apenas aqui)
-    setScrollTarget('events', window.scrollY);
+    setScrollTarget('services', window.scrollY);
 
     // navega para detalhes normalmente
     navigate(targetUrl);
@@ -51,7 +50,7 @@ export default function Eventos() {
   if (loading) {
     return (
       <section
-        id="events"
+        id="services"
         className="w-full py-12 px-4 md:px-8 bg-gray-50 mb-6"
       >
         <div className="max-w-3xl mx-auto text-center">
@@ -59,10 +58,10 @@ export default function Eventos() {
             variant="h1"
             className="text-3xl md:text-5xl font-bold text-primary"
           >
-            Eventos Disponíveis
+            Serviços Disponíveis
           </Typography>
           <Typography variant="p" className="text-gray-600 mt-2">
-            Carregando eventos...
+            Carregando serviços...
           </Typography>
         </div>
       </section>
@@ -72,7 +71,7 @@ export default function Eventos() {
   if (error) {
     return (
       <section
-        id="events"
+        id="services"
         className="w-full py-12 px-4 md:px-8 bg-gray-50 mb-6"
       >
         <div className="max-w-3xl mx-auto text-center">
@@ -80,7 +79,7 @@ export default function Eventos() {
             variant="h1"
             className="text-3xl md:text-5xl font-bold text-primary"
           >
-            Eventos Disponíveis
+            Serviços Disponíveis
           </Typography>
           <Typography variant="p" className="text-red-600 mt-2">
             {error}
@@ -90,10 +89,10 @@ export default function Eventos() {
     );
   }
 
-  if (events.length === 0) {
+  if (services.length === 0) {
     return (
       <section
-        id="events"
+        id="services"
         className="w-full py-12 px-4 md:px-8 bg-gray-50 mb-6"
       >
         <div className="max-w-3xl mx-auto text-center">
@@ -101,10 +100,10 @@ export default function Eventos() {
             variant="h1"
             className="text-3xl md:text-5xl font-bold text-primary"
           >
-            Eventos Disponíveis
+            Serviços Disponíveis
           </Typography>
           <Typography variant="p" className="text-gray-600 mt-2">
-            Nenhum evento disponível no momento.
+            Nenhum serviço disponível no momento.
           </Typography>
         </div>
       </section>
@@ -112,17 +111,20 @@ export default function Eventos() {
   }
 
   return (
-    <section id="events" className="w-full py-12 px-4 md:px-8 bg-white mb-6">
+    <section
+      id="services"
+      className="w-full py-12 px-4 md:px-8 bg-gray-50 mb-6"
+    >
       <div className="max-w-3xl mx-auto text-center">
         <Typography
           variant="h1"
           className="text-3xl md:text-5xl font-bold text-primary"
         >
-          Eventos
+          Serviços Disponíveis
         </Typography>
         <Typography variant="p" className="text-gray-600 mt-2">
-          Participe e faça a diferença! Acompanhe nossos eventos e junte-se às
-          ações que transformam vidas.
+          Conheça os serviços oferecidos pela nossa rede social solidária,
+          desenvolvidos para apoiar e transformar vidas.
         </Typography>
       </div>
 
@@ -151,58 +153,51 @@ export default function Eventos() {
             },
           }}
         >
-          {events.map((event) => (
-            <SwiperSlide key={event.id}>
+          {services.map((service) => (
+            <SwiperSlide key={service.id}>
               <div className="bg-white border rounded-lg shadow-md overflow-hidden flex flex-col h-full">
+                {/* Container da imagem com altura fixa mas proporção preservada */}
                 <div
                   className={`w-full aspect-[4/3] flex items-center justify-center rounded-t-lg overflow-hidden ${
-                    !event.image || imageErrors[String(event.id)]
+                    !service.image || imageErrors[String(service.id)]
                       ? 'bg-[#f0f9ff]'
                       : 'bg-gradient-to-b from-gray-50 to-gray-100'
                   }`}
                 >
                   <img
                     src={
-                      event.image && !imageErrors[String(event.id)]
-                        ? resolveImageUrl(event.image)
-                        : '/img/evento-solidario.png'
+                      service.image && !imageErrors[String(service.id)]
+                        ? resolveImageUrl(service.image)
+                        : '/img/servico-solidario2.png'
                     }
-                    alt={event.title}
+                    alt={service.title}
                     className="object-contain w-full h-full transition-transform duration-300"
-                    onError={() => handleImageError(String(event.id))}
+                    onError={() => handleImageError(String(service.id))}
                   />
                 </div>
 
-                {/* Conteúdo */}
+                {/* Resto do conteúdo permanece igual */}
                 <div className="p-3 flex flex-col flex-1">
                   <div className="h-10 mb-1">
                     <Typography
                       variant="h3"
                       className="text-base text-gray-900 line-clamp-2"
                     >
-                      {event.title}
+                      {service.title}
                     </Typography>
                   </div>
 
                   <div className="h-14 mb-2">
                     <Typography
                       variant="p"
-                      className="text-sm text-gray-600 space-y-1"
+                      className="text-sm text-gray-600 line-clamp-2"
                     >
-                      <span>
-                        <strong>Data:</strong> {formatDateBR(event.date)}
-                      </span>
-                      <span className="flex">
-                        <strong className="flex-shrink-0">Local:</strong>
-                        <span className="line-clamp-2 ml-1">
-                          {event.location}
-                        </span>
-                      </span>
+                      {service.description}
                     </Typography>
                   </div>
 
                   <div className="mt-2">
-                    <CardButton onClick={() => handleEventClick(event)}>
+                    <CardButton onClick={() => handleServiceClick(service)}>
                       Saiba Mais
                     </CardButton>
                   </div>
@@ -212,7 +207,7 @@ export default function Eventos() {
           ))}
         </Swiper>
 
-        {/* Botões de navegação */}
+        {/* Botões de navegação - posicionados fora do card */}
         <div className="swiper-button-prev text-3xl absolute left-0 top-1/2 transform -translate-y-1/2 text-gray-700 hover:text-blue-600"></div>
         <div className="swiper-button-next text-3xl absolute right-0 top-1/2 transform -translate-y-1/2 text-gray-700 hover:text-blue-600"></div>
       </div>
