+++ src/components/ui/EngagementCounters.tsx
@@ -1,6 +1,5 @@
 // components/ui/EngagementCounters.tsx
 import React from 'react';
-import { FaHeart, FaComment, FaShare, FaUserCheck } from 'react-icons/fa';
 
 interface EngagementCountersProps {
   likesCount?: number;
@@ -8,6 +7,7 @@ interface EngagementCountersProps {
   sharesCount?: number;
   attendanceCount?: number;
   categoryId: number;
+  compact?: boolean;
 }
 
 export const EngagementCounters: React.FC<EngagementCountersProps> = ({
@@ -16,43 +16,80 @@ export const EngagementCounters: React.FC<EngagementCountersProps> = ({
   sharesCount,
   attendanceCount,
   categoryId,
+  compact = false,
 }) => {
   if (!likesCount && !commentsCount && !sharesCount && !attendanceCount) {
     return null;
   }
 
+  // ✅ VERSÃO COMPACTA COM TEXTO PURO - Sem ícones
+  if (compact) {
+    return (
+      <div className="flex flex-wrap items-center gap-3 text-xs text-gray-600">
+        {/* Curtidas */}
+        {likesCount !== undefined && likesCount > 0 && (
+          <span className="whitespace-nowrap">
+            <strong>{likesCount}</strong>{' '}
+            {likesCount === 1 ? 'curtida' : 'curtidas'}
+          </span>
+        )}
+
+        {/* Comentários */}
+        {commentsCount !== undefined && commentsCount > 0 && (
+          <span className="whitespace-nowrap">
+            <strong>{commentsCount}</strong>{' '}
+            {commentsCount === 1 ? 'comentário' : 'comentários'}
+          </span>
+        )}
+
+        {/* Compartilhamentos */}
+        {sharesCount !== undefined && sharesCount > 0 && (
+          <span className="whitespace-nowrap">
+            <strong>{sharesCount}</strong>{' '}
+            {sharesCount === 1 ? 'compart.' : 'compart.'}
+          </span>
+        )}
+
+        {/* Presenças */}
+        {categoryId === 8 &&
+          attendanceCount !== undefined &&
+          attendanceCount > 0 && (
+            <span className="whitespace-nowrap">
+              <strong>{attendanceCount}</strong>{' '}
+              {attendanceCount === 1 ? 'confirmado' : 'confirmados'}
+            </span>
+          )}
+      </div>
+    );
+  }
+
+  // ✅ VERSÃO ORIGINAL (como fallback para outros usos)
   return (
     <div className="flex items-center gap-4 text-xs text-gray-500 border-t pt-2 mt-2">
       {/* Curtidas */}
       {likesCount !== undefined && likesCount > 0 && (
-        <div
-          className="flex items-center gap-1.5"
-          title={`${likesCount} curtidas`}
-        >
-          <FaHeart className="w-3.5 h-3.5" />
-          <span>{likesCount}</span>
+        <div className="flex items-center gap-1.5">
+          <span>
+            <strong>{likesCount}</strong> curtidas
+          </span>
         </div>
       )}
 
       {/* Comentários */}
       {commentsCount !== undefined && commentsCount > 0 && (
-        <div
-          className="flex items-center gap-1.5"
-          title={`${commentsCount} comentários`}
-        >
-          <FaComment className="w-3.5 h-3.5" />
-          <span>{commentsCount}</span>
+        <div className="flex items-center gap-1.5">
+          <span>
+            <strong>{commentsCount}</strong> comentários
+          </span>
         </div>
       )}
 
       {/* Compartilhamentos */}
       {sharesCount !== undefined && sharesCount > 0 && (
-        <div
-          className="flex items-center gap-1.5"
-          title={`${sharesCount} compartilhamentos`}
-        >
-          <FaShare className="w-3.5 h-3.5" />
-          <span>{sharesCount}</span>
+        <div className="flex items-center gap-1.5">
+          <span>
+            <strong>{sharesCount}</strong> compart.
+          </span>
         </div>
       )}
 
@@ -60,12 +97,10 @@ export const EngagementCounters: React.FC<EngagementCountersProps> = ({
       {categoryId === 8 &&
         attendanceCount !== undefined &&
         attendanceCount > 0 && (
-          <div
-            className="flex items-center gap-1.5"
-            title={`${attendanceCount} confirmados`}
-          >
-            <FaUserCheck className="w-3.5 h-3.5" />
-            <span>{attendanceCount}</span>
+          <div className="flex items-center gap-1.5">
+            <span>
+              <strong>{attendanceCount}</strong> confirmados
+            </span>
           </div>
         )}
     </div>
