+++ src/pages/ProfileView.tsx
@@ -8,9 +8,9 @@ import BackButton from '../components/ui/BackButton';
 import { useProfile } from '../hooks/useProfile';
 import { useUserProfile } from '../hooks/useUserProfile';
 import { useProfilePosts } from '../hooks/useProfilePosts';
-import PostCard from '../components/PostCard';
-import PostModal from '../components/PostModal';
-import ShareModal from '../components/ShareModal';
+import PostCard from '../components/posts/PostCard';
+import PostModal from '../components/posts/PostModal';
+import ShareModal from '../components/posts/ShareModal';
 import EditPostModal from '../components/posts/EditPostModal';
 import ShareEditModal from '../components/posts/ShareEditModal';
 import { usePostStore } from '../stores/postStore';
